Cambios menores
A ver no es que haya cambiado mucho la cosa, salvo una excepción: que el módulo esté en formato double y no int, porque si no me sale 0 el módulo y la división con x e y hace dé infinito (x/0).

+++ HolaSDL/Vector2D.cpp
@@ -8,7 +8,7 @@
 
 Vector2D Vector2D::normalizeVector()
 {
-	int mod = (this->x * this->x) + (this->y * this->y);
+	double mod = (this->x * this->x) + (this->y * this->y);
 	mod = sqrt(mod);
 	return Vector2D(this->x / mod, this->y / mod);
 }
